replaced uints with size_t

+++ src/kernels/MIOpenDropoutHIP.cpp
@@ -70,10 +70,10 @@ extern "C" __global__ void
 InitKernelStateHIP(rocrand_state_xorwow* state, ulong prng_seed, ulong states_num)
 {
     // Get the index of the current element
-    uint index  = blockIdx.x * blockDim.x + threadIdx.x;
-    uint stride = blockDim.x * gridDim.x;
+    size_t index  = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t stride = blockDim.x * gridDim.x;
 
-    for(uint gid = index; gid < states_num; gid += stride)
+    for(size_t gid = index; gid < states_num; gid += stride)
     {
         rocrand_state_xorwow state_gid;
         rocrand_init(prng_seed, gid, 0ULL, &state_gid);
@@ -120,31 +120,31 @@ __forceinline__ __device__ void dropoutfw(const rocrand_state_xorwow* state,
                                           int in_str2,
                                           int in_str3,
                                           uchar* reserveSpace,
-                                          uint total_work,
-                                          uint in_offset,
-                                          uint out_offset,
-                                          uint rsvsp_offset)
+                                          size_t total_work,
+                                          size_t in_offset,
+                                          size_t out_offset,
+                                          size_t rsvsp_offset)
 {
     F dat_blk[RD_BLCK];     // Register space to read the input data
     uchar is_kept[RD_BLCK]; // Register space to store the mask for the dropout
 
-    uint sid = threadIdx.x + blockIdx.x * blockDim.x;
+    size_t sid = threadIdx.x + blockIdx.x * blockDim.x;
     rocrand_state_xorwow cur_state; // Read the state of the current thread
     cur_state = state[sid];
 
-    for(uint gid = threadIdx.x + blockIdx.x * blockDim.x; gid < total_work;
+    for(size_t gid = threadIdx.x + blockIdx.x * blockDim.x; gid < total_work;
         gid += blockDim.x * gridDim.x)
     {
-        uint i0    = gid / (dim1 * dim2 * dim3 * dim4);
-        uint i1    = (gid / (dim2 * dim3 * dim4)) % dim1;
-        uint i2    = (gid / (dim3 * dim4)) % dim2;
-        uint i3    = (gid / dim4) % dim3;
-        uint i4    = gid % dim4;
-        uint i4_rd = i4 / RD_BLCK;
+        size_t i0    = gid / (dim1 * dim2 * dim3 * dim4);
+        size_t i1    = (gid / (dim2 * dim3 * dim4)) % dim1;
+        size_t i2    = (gid / (dim3 * dim4)) % dim2;
+        size_t i3    = (gid / dim4) % dim3;
+        size_t i4    = gid % dim4;
+        size_t i4_rd = i4 / RD_BLCK;
 
-        uint x_idx = i0 * in_str0 + i1 * in_str1 + i2 * in_str2 + i3 * in_str3 +
+        size_t x_idx = i0 * in_str0 + i1 * in_str1 + i2 * in_str2 + i3 * in_str3 +
                      i4_rd * RD_BLCK; // Calculate the index of the input tensor
-        uint y_idx = i0 * out_str0 + i1 * out_str1 + i2 * out_str2 + i3 * out_str3 +
+        size_t y_idx = i0 * out_str0 + i1 * out_str1 + i2 * out_str2 + i3 * out_str3 +
                      i4_rd * RD_BLCK; // Calculate the index of the output tensor
 
         *(reinterpret_cast<T*>(dat_blk)) = *(reinterpret_cast<const T*>(
@@ -152,7 +152,7 @@ __forceinline__ __device__ void dropoutfw(const rocrand_state_xorwow* state,
 
         if constexpr(!MASK) // If MASK is not enabled then generate the mask for dropout
         {
-            for(int i = 0; i < RD_BLCK; ++i)
+            for(size_t i = 0; i < RD_BLCK; ++i)
             {
                 is_kept[i] =
                     static_cast<uchar>(prng::xorwow_uniform(&cur_state) >
@@ -173,7 +173,7 @@ __forceinline__ __device__ void dropoutfw(const rocrand_state_xorwow* state,
                 reserveSpace + rsvsp_offset + gid - i4 + i4_rd * RD_BLCK));
         }
         // Apply the mask to the data and scale it with the scale factor.
-        for(int i = 0; i < RD_BLCK; ++i)
+        for(size_t i = 0; i < RD_BLCK; ++i)
         {
             dat_blk[i] = static_cast<bool>(is_kept[i]) ? dat_blk[i] * (F)scale : (F)0;
         }
@@ -200,10 +200,10 @@ extern "C" __global__ void DropoutFW(const rocrand_state_xorwow* state,
                                      int in_str2,
                                      int in_str3,
                                      uchar* reserveSpace,
-                                     uint total_work,
-                                     uint in_offset,
-                                     uint out_offset,
-                                     uint rsvsp_offset)
+                                     size_t total_work,
+                                     size_t in_offset,
+                                     size_t out_offset,
+                                     size_t rsvsp_offset)
 {
 
     dropoutfw<_FLOAT, READ_DAT_TYPE, READ_BOOL_TYPE, USE_MASK, USE_RSVSP>(state,
@@ -264,31 +264,31 @@ __forceinline__ __device__ void dropoutbw(const rocrand_state_xorwow* state,
                                           int in_str2,
                                           int in_str3,
                                           uchar* reserveSpace,
-                                          uint total_work,
-                                          uint in_offset,
-                                          uint out_offset,
-                                          uint rsvsp_offset)
+                                          size_t total_work,
+                                          size_t in_offset,
+                                          size_t out_offset,
+                                          size_t rsvsp_offset)
 {
     F dat_blk[RD_BLCK];     // Read in the input data into the register space
     uchar is_kept[RD_BLCK]; // is_kept holds the mask for the dropout
 
-    uint sid = threadIdx.x + blockIdx.x * blockDim.x;
+    size_t sid = threadIdx.x + blockIdx.x * blockDim.x;
     rocrand_state_xorwow cur_state;
     cur_state = state[sid];
 
-    for(uint gid = threadIdx.x + blockIdx.x * blockDim.x; gid < total_work;
+    for(size_t gid = threadIdx.x + blockIdx.x * blockDim.x; gid < total_work;
         gid += blockDim.x * gridDim.x)
     {
-        uint i0    = gid / (dim1 * dim2 * dim3 * dim4);
-        uint i1    = (gid / (dim2 * dim3 * dim4)) % dim1;
-        uint i2    = (gid / (dim3 * dim4)) % dim2;
-        uint i3    = (gid / dim4) % dim3;
-        uint i4    = gid % dim4;
-        uint i4_rd = i4 / RD_BLCK;
+        size_t i0    = gid / (dim1 * dim2 * dim3 * dim4);
+        size_t i1    = (gid / (dim2 * dim3 * dim4)) % dim1;
+        size_t i2    = (gid / (dim3 * dim4)) % dim2;
+        size_t i3    = (gid / dim4) % dim3;
+        size_t i4    = gid % dim4;
+        size_t i4_rd = i4 / RD_BLCK;
 
-        uint x_idx = i0 * in_str0 + i1 * in_str1 + i2 * in_str2 + i3 * in_str3 +
+        size_t x_idx = i0 * in_str0 + i1 * in_str1 + i2 * in_str2 + i3 * in_str3 +
                      i4_rd * RD_BLCK; // Calculate the index of the output tensor
-        uint y_idx = i0 * out_str0 + i1 * out_str1 + i2 * out_str2 + i3 * out_str3 +
+        size_t y_idx = i0 * out_str0 + i1 * out_str1 + i2 * out_str2 + i3 * out_str3 +
                      i4_rd * RD_BLCK; // Calculate the index of the input tensor
 
         // Read RD_BLCK number of _FLOAT data from y and store it in dat_blk
@@ -298,7 +298,7 @@ __forceinline__ __device__ void dropoutbw(const rocrand_state_xorwow* state,
         if constexpr(PRNG)
         {
             // Generate mask for RD_BLCK number of elements when it is a vectorized read.
-            for(int i = 0; i < RD_BLCK; ++i)
+            for(size_t i = 0; i < RD_BLCK; ++i)
             {
 
                 is_kept[i] = static_cast<uchar>(prng::xorwow_uniform(&cur_state) > dropout);
@@ -312,7 +312,7 @@ __forceinline__ __device__ void dropoutbw(const rocrand_state_xorwow* state,
         }
 
         // Iterate over the RD_BLCK number of elements and apply the mask to the data
-        for(int i = 0; i < RD_BLCK; ++i)
+        for(size_t i = 0; i < RD_BLCK; ++i)
         {
             // If the element is retained then scale it with the scale factor
             dat_blk[i] = static_cast<bool>(is_kept[i]) ? dat_blk[i] * (F)scale : (F)0;
@@ -341,10 +341,10 @@ extern "C" __global__ void DropoutBW(const rocrand_state_xorwow* state,
                                      int in_str2,
                                      int in_str3,
                                      uchar* reserveSpace,
-                                     uint total_work,
-                                     uint in_offset,
-                                     uint out_offset,
-                                     uint rsvsp_offset)
+                                     size_t total_work,
+                                     size_t in_offset,
+                                     size_t out_offset,
+                                     size_t rsvsp_offset)
 {
 
     dropoutbw<_FLOAT, READ_DAT_TYPE, READ_BOOL_TYPE, USE_PRNG>(state,
